🏗️ refactor: decouple agent-specific usage logic into a modular registry and shared utility layer.
